Add new PaymentException to throw from both Adyen and Shopware processes
Shopware pay() and finalize() methods are using different exception
types to indicate if there was an error with the payment during pay or
finalize. We want to handle them with the same processor so a standard
exception needs to be thrown and when caught the proper shopware
exception needs to be thrown.