[PW-2691] Handle response back from an issuer page (#25)
* Add new PaymentException to throw from both Adyen and Shopware processes

Shopware pay() and finalize() methods are using different exception
types to indicate if there was an error with the payment during pay or
finalize. We want to handle them with the same processor so a standard
exception needs to be thrown and when caught the proper shopware
exception needs to be thrown.

* Make doPaymentDetails to handle responses from API and async methods too

Accept any prepared array details and orderNumber to validate against
the paymentDetails Adyen endpoint. This function should not be
responsible preparing the payload only to send the request and return
the handled response from the handlePaymentResponse()

* Validate stateData before the paymentResponseHandler handles the payload

The paymentResponseHandler->handleAdyenApis is used for verifying API
and asynchronous responses as well. Therefore at first validate and
prepare the stateData from the payment-details plugin endpoint

* Adjust CardsPaymentMethodHandler to the new handler and exception

* Adjust the PaymentResponseHandler.php to the new exception

* Implement finalize() for cards

* phpcs fixes and formatting

* Remove duplicated setCustomFields

* Fix wrong service import

* Remove unused service from ResultHandler.php

* Fix validation and throwing exception in CardsPaymentMethodHandler.php

* Remove unused $paymentResponseHandlerResult and other use classes

* Remove order_transaction.repository from service PaymentResponseHandler

* Accept only payload from the query parameters when handling result

Do not pass all the request parameters to the payment details in case of
a redirect result validation only the payload

* Update src/Handlers/ResultHandler.php

Co-authored-by: Marcos Garcia <marcos.asgarcia@gmail.com>

* Throw PaymentException in ResultHandler instead of Async exception

Since the ResultHandler.php can be used by other methods and not just
by finalize, make sure to throw the more generic PaymentException and
only throw AsyncPaymentFinalizeException from the finalize()

* Add missing service for Cards handler

* Generate payment-status url and pass orderId to the request

Co-authored-by: Marcos Garcia <marcos.asgarcia@gmail.com>